PubSub/requirements.txt
apturl==0.5.2
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==3.2.0
black==22.10.0
blessed==1.19.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cfgv==3.3.1
chardet==4.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-picklefield==3.1
django-q==1.3.9
django-simple-history==3.1.1
djangorestframework==3.14.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
httplib2==0.20.2
identify==2.5.6
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
nodeenv==1.7.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pathspec==0.10.1
pexpect==4.8.0
Pillow==9.0.1
pipenv @ git+https://github.com/pypa/pipenv.git@cb9d87ad05830449743cbc3ffa09ebbc7cca602d
platformdirs==2.5.2
pre-commit==2.20.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymemcache==4.0.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python3-memcached==1.51
pytz==2022.4
pyxdg==0.27
PyYAML==6.0
redis==3.5.3
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
toml==0.10.2
tomli==2.0.1
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.16.5
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.5
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
